# portfolio/public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "Minimal"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
...
  "paths": {
...
      "/add-photos": {
          "post": {
              "tags": ["Photo"],
              "summary": "Add multiple photos (up to 10) for the authenticated user",
              "operationId": "addPhotos",
              "consumes": ["application/json"],
              "produces": ["application/json"],
              "parameters": [
                  {
                      "name": "Authorization",
                      "in": "header",
                      "description": "Bearer token for authentication",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "name": "body",
                      "in": "body",
                      "description": "Array of photos to be added",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "photos": {
                                  "type": "array",
                                  "items": {
                                      "type": "object",
                                      "properties": {
                                          "title": { "type": "string" },
                                          "description": { "type": "string" },
                                          "location": { "type": "string" },
                                          "image_url": { "type": "string" },
                                          "category_id": { "type": "integer" },
                                          "privacy_status": { "type": "string" },
                                          "tags": { "type": "string" }
                                      },
                                      "required": ["title", "image_url", "category_id", "privacy_status"]
                                  }
                              }
                          }
                      }
                  }
              ],
              "responses": {
                  "201": {
                      "description": "Photos added successfully",
                      "schema": {
                          "type": "object",
                          "properties": {
                              "message": { "type": "string" }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation error",
                      "schema": {
                          "type": "object",
                          "properties": {
                              "errors": {
                                  "type": "object",
                                  "additionalProperties": { "type": "string" }
                              }
                          }
                      }
                  },
                  "500": {
                      "description": "Internal server error"
                  }
              }
          }
      },
...
      "/notifications": {
          "get": {
              "tags": ["Notification"],
              "summary": "Get all notifications for the current user",
              "operationId": "getUserNotifications",
              "produces": ["application/json"],
              "parameters": [
                  {
                      "name": "Authorization",
                      "in": "header",
                      "description": "Bearer token for authentication",
                      "required": true,
                      "type": "string"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Notifications retrieved successfully",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/Notification"
                          }
                      }
                  },
                  "401": {
                      "description": "Unauthorized"
                  },
                  "500": {
                      "description": "Internal server error"
                  }
              }
          }
      },
...
},
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ]
}